pyproject.toml
[project]
name = "eink-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gpiozero>=2.0.1",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "spidev>=3.8",
]
